requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
pillow==10.4.0
pytesseract==0.3.13
python-pptx==1.0.2
unstructured[all-docs]==0.15.13
openai==1.51.0
httpx==0.27.2